i think cvss31 is working. now its time to make it pretty

+++ src/types.zig
@@ -1,7 +1,7 @@
 // const CVSS4_HEADER = "CVSS:4.0";
 
 pub const CvssScore = extern struct {
-    score: f16,
+    score: f32,
     level: CVSS_LEVEL,
 };
